divide service and handler logic
